// src/app/application-name/shared/models/dto/user-dto.interface.ts
export interface UserDto {
  id?: number;

  name?: string;

  firstname?: string;

  title?: string;

  function?: string;

  fax?: string;

  phone?: string;

  mobile?: string;

  email?: string;

  language?: string;

  ynUcAcceptedSimpl?: string;

  lastLovDate?: Date;

  lastParamDate?: Date;

}